Major refactor to remove circular dependencies (#730)
* Major refactor to remove circular dependencies.

Key changes:

* break out BaseTraitHandler, TraitHandler, TraitType, CTrait, and TraitFactory
  into their own submodules.
* take the various trait converters/trait casters etc. and put them into their
  own submodule; plus simplify code via use of an informal `as_ctrait`
  interface so that we don't need as many imports.
* remove the dependency of List TraitType on items_event
* assorted other reshuffles and clean-up

As much as possible, if a name is available in a module, I've left it imported
from its new location for backward compatibility.

* Always _save_ after fixing a merge.

* Fixes for documentation generation.

* Add tests for trait converters.
