+++ docs/source/traits_api_reference/base_trait_handler.rst
@@ -1,0 +1,10 @@
+:mod:`base_trait_handler` Module
+================================
+
+.. automodule:: traits.base_trait_handler
+    :no-members:
+
+Classes
+-------
+
+.. autoclass:: BaseTraitHandler
