+++ docs/source/traits_api_reference/ctrait.rst
@@ -1,0 +1,12 @@
+:mod:`ctrait` Module
+====================
+
+.. automodule:: traits.ctrait
+    :no-members:
+
+.. currentmodule:: traits.ctrait
+
+Classes
+-------
+
+.. autoclass:: CTrait
