+++ docs/source/traits_api_reference/index.rst
@@ -9,19 +9,26 @@ Traits core
 
     traits
     has_traits
-    interface_checker
+
+    base_trait_handler
     constants
-    trait_base
-    trait_errors
-    trait_handlers
-    trait_numeric
-    trait_types
-    traits_listener
-    trait_notifiers
+    ctrait
     editor_factories
+    interface_checker
+    trait_base
+    trait_converters
     trait_dict_object
+    trait_errors
+    trait_factory
+    trait_handler
+    trait_handlers
     trait_list_object
+    trait_numeric
+    trait_type
+    trait_types
+    trait_notifiers
     trait_set_object
+    traits_listener
     version
 
 Subpackages
