+++ docs/source/traits_api_reference/trait_base.rst
@@ -15,6 +15,8 @@ Classes
 
 .. autodata:: Self
 
+.. autoclass:: HandleWeakRef
+
 Functions
 ---------
