+++ docs/source/traits_api_reference/trait_converters.rst
@@ -1,0 +1,21 @@
+:mod:`trait_converters` Module
+==============================
+
+.. automodule:: traits.trait_converters
+    :no-members:
+
+.. currentmodule:: traits.trait_converters
+
+
+Functions
+---------
+
+.. autofunction:: check_trait
+
+.. autofunction:: trait_cast
+
+.. autofunction:: trait_from
+
+.. autofunction:: trait_for
+
+.. autofunction:: mapped_trait_for
