+++ docs/source/traits_api_reference/trait_factory.rst
@@ -1,0 +1,20 @@
+:mod:`trait_factory` Module
+===========================
+
+.. automodule:: traits.trait_factory
+    :no-members:
+
+.. currentmodule:: traits.trait_factory
+
+Classes
+-------
+
+.. autoclass:: TraitFactory
+
+.. autoclass:: TraitImportError
+
+
+Functions
+---------
+
+.. autofunction:: trait_factory
