+++ docs/source/traits_api_reference/trait_handler.rst
@@ -1,0 +1,10 @@
+:mod:`trait_handler` Module
+============================
+
+.. automodule:: traits.trait_handler
+    :no-members:
+
+Classes
+-------
+
+.. autoclass:: TraitHandler
