+++ docs/source/traits_api_reference/trait_type.rst
@@ -1,0 +1,22 @@
+:mod:`trait_type` Module
+========================
+
+.. automodule:: traits.trait_type
+    :no-members:
+
+Classes
+-------
+
+.. autoclass:: NoDefaultSpecified
+
+.. autoclass:: TraitType
+
+Private Functions
+-----------------
+.. autofunction:: traits.trait_type._infer_default_value_type
+
+.. autofunction:: traits.trait_type._arg_count
+
+.. autofunction:: traits.trait_type._write_only
+
+.. autofunction:: traits.trait_type._read_only
