+++ traits/tests/test_ctraits.py
@@ -16,7 +16,7 @@ import warnings
 from traits.constants import (
     ComparisonMode, DefaultValue, TraitKind, MAXIMUM_DEFAULT_VALUE_TYPE
 )
-from traits.traits import CTrait
+from traits.ctrait import CTrait
 
 
 def getter():
