+++ traits/tests/test_has_traits.py
@@ -10,7 +10,8 @@ from traits.has_traits import (
     InstanceTraits,
     HasTraits,
 )
-from traits.traits import CTrait, ForwardProperty, generic_trait
+from traits.ctrait import CTrait
+from traits.traits import ForwardProperty, generic_trait
 from traits.trait_types import Event, Float, Instance, Int
