+++ traits/trait_base.py
@@ -28,6 +28,7 @@ import os
 import sys
 from os import getcwd
 from os.path import dirname, exists, join
+from weakref import ref
 
 from .etsconfig.api import ETSConfig
 from .constants import ValidateTrait
@@ -42,6 +43,8 @@ enumerate = enumerate
 SequenceTypes = (list, tuple)
 
 ComplexTypes = (float, int)
+
+RangeTypes = (int, float)
 
 TypeTypes = (
     str,
@@ -394,6 +397,28 @@ def xsetattr(object, xname, value):
         object = getattr(object, name)
 
     setattr(object, names[-1], value)
+
+# -------------------------------------------------------------------------------
+# Helpers for weak references
+# -------------------------------------------------------------------------------
+
+def _make_value_freed_callback(object_ref, name):
+    def _value_freed(value_ref):
+        object = object_ref()
+        if object is not None:
+            object.trait_property_changed(name, Undefined, None)
+
+    return _value_freed
+
+
+class HandleWeakRef(object):
+    def __init__(self, object, name, value):
+        object_ref = ref(object)
+        _value_freed = _make_value_freed_callback(object_ref, name)
+        self.object = object_ref
+        self.name = name
+        self.value = ref(value, _value_freed)
+
 
 
 # -------------------------------------------------------------------------------
