+++ traits/trait_handlers.py
@@ -18,8 +18,7 @@
 # ------------------------------------------------------------------------------
 
 """
-Defines the BaseTraitHandler class and a standard set of BaseTraitHandler
-subclasses for use with the Traits package.
+Defines a standard set of TraitHandler subclasses.
 
 A trait handler mediates the assignment of values to object traits. It
 verifies (via its validate() method) that a specified value is consistent
@@ -35,9 +34,10 @@ from importlib import import_module
 import re
 import sys
 from types import FunctionType, MethodType
-from weakref import ref
 
 from .constants import DefaultValue, TraitKind, ValidateTrait
+from .base_trait_handler import BaseTraitHandler
+from .trait_type import TraitType, trait_types
 from .trait_base import (
     strx,
     SequenceTypes,
@@ -47,16 +47,15 @@ from .trait_base import (
     TraitsCache,
     class_of,
     Missing,
+    RangeTypes,
     Self,
 )
 from .trait_errors import TraitError
 from .trait_dict_object import TraitDictEvent, TraitDictObject
+from .trait_converters import trait_from
+from .trait_handler import TraitHandler
 from .trait_list_object import TraitListEvent, TraitListObject
 from .trait_set_object import TraitSetEvent, TraitSetObject
-
-
-# Patched by 'traits.py' once class is defined!
-Trait = Event = None
 
 # Set up a logger:
 import logging
@@ -67,50 +66,12 @@ logger = logging.getLogger(__name__)
 #  Constants:
 # -------------------------------------------------------------------------------
 
-RangeTypes = (int, float)
-
 CallableTypes = (FunctionType, MethodType)
 
-# Mapping from trait metadata 'type' to CTrait 'type':
-trait_types = {"python": 1, "event": 2}
 
 # -------------------------------------------------------------------------------
-#  Forward references:
+#  Private functions:
 # -------------------------------------------------------------------------------
-
-trait_from = None  # Patched by 'traits.py' when real 'trait_from' is defined
-
-# -------------------------------------------------------------------------------
-#  Returns the correct argument count for a specified function or method:
-# -------------------------------------------------------------------------------
-
-
-def _arg_count(func):
-    """ Returns the correct argument count for a specified function or method.
-    """
-    if (type(func) is MethodType) and (func.__self__ is not None):
-        return func.__code__.co_argcount - 1
-    return func.__code__.co_argcount
-
-
-# -------------------------------------------------------------------------------
-#  Property error handling functions:
-# -------------------------------------------------------------------------------
-
-
-def _write_only(object, name):
-    raise TraitError(
-        "The '%s' trait of %s instance is 'write only'."
-        % (name, class_of(object))
-    )
-
-
-def _read_only(object, name, value):
-    raise TraitError(
-        "The '%s' trait of %s instance is 'read only'."
-        % (name, class_of(object))
-    )
-
 
 def _undefined_get(object, name):
     raise TraitError(
@@ -124,593 +85,6 @@ def _undefined_get(object, name):
 
 def _undefined_set(object, name, value):
     _undefined_get(object, name)
-
-
-def _infer_default_value_type(default_value):
-    """
-    Figure out the appropriate default value type given a default value.
-    """
-    if default_value is Missing:
-        return DefaultValue.missing
-    elif default_value is Self:
-        return DefaultValue.object
-    elif isinstance(default_value, TraitListObject):
-        return DefaultValue.trait_list_object
-    elif isinstance(default_value, TraitDictObject):
-        return DefaultValue.trait_dict_object
-    elif isinstance(default_value, TraitSetObject):
-        return DefaultValue.trait_set_object
-    elif isinstance(default_value, list):
-        return DefaultValue.list_copy
-    elif isinstance(default_value, dict):
-        return DefaultValue.dict_copy
-    else:
-        return DefaultValue.constant
-
-
-# -------------------------------------------------------------------------------
-#  'BaseTraitHandler' class (base class for all user defined traits and trait
-#  handlers):
-# -------------------------------------------------------------------------------
-
-
-class BaseTraitHandler(object):
-    """ The task of this class and its subclasses is to verify the correctness
-    of values assigned to object trait attributes.
-
-    This class is an alternative to trait validator functions. A trait handler
-    has several advantages over a trait validator function, due to being an
-    object:
-
-        * Trait handlers have constructors and state. Therefore, you can use
-          them to create *parametrized types*.
-        * Trait handlers can have multiple methods, whereas validator functions
-          can have only one callable interface. This feature allows more
-          flexibility in their implementation, and allows them to handle a
-          wider range of cases, such as interactions with other components.
-    """
-
-    default_value_type = DefaultValue.unspecified
-    has_items = False
-    is_mapped = False
-    editor = None
-    info_text = "a legal value"
-
-    def is_valid(self, object, name, value):
-        try:
-            validate = self.validate
-            try:
-                validate(object, name, value)
-                return True
-            except:
-                return False
-        except:
-            return True
-
-    def error(self, object, name, value):
-        """Raises a TraitError exception.
-
-        Parameters
-        ----------
-        object : object
-            The object whose attribute is being assigned.
-        name : str
-            The name of the attribute being assigned.
-        value : object
-            The proposed new value for the attribute.
-
-        Description
-        -----------
-        This method is called by the validate() method when an assigned value
-        is not valid. Raising a TraitError exception either notifies the user of
-        the problem, or, in the case of compound traits, provides a chance for
-        another trait handler to handle to validate the value.
-        """
-        raise TraitError(
-            object, name, self.full_info(object, name, value), value
-        )
-
-    def full_info(self, object, name, value):
-        """Returns a string describing the type of value accepted by the
-        trait handler.
-
-        Parameters
-        ----------
-        object : object
-            The object whose attribute is being assigned.
-        name : str
-            The name of the attribute being assigned.
-        value :
-            The proposed new value for the attribute.
-
-        Description
-        -----------
-        The string should be a phrase describing the type defined by the
-        TraitHandler subclass, rather than a complete sentence. For example, use
-        the phrase, "a square sprocket" instead of the sentence, "The value must
-        be a square sprocket." The value returned by full_info() is combined
-        with other information whenever an error occurs and therefore makes more
-        sense to the user if the result is a phrase. The full_info() method is
-        similar in purpose and use to the **info** attribute of a validator
-        function.
-
-        Note that the result can include information specific to the particular
-        trait handler instance. If the full_info() method is not overridden,
-        the default method returns the value of calling the info() method.
-        """
-        return self.info()
-
-    def info(self):
-        """Must return a string describing the type of value accepted by the
-        trait handler.
-
-        The string should be a phrase describing the type defined by the
-        TraitHandler subclass, rather than a complete sentence. For example, use
-        the phrase, "a square sprocket" instead of the sentence, "The value must
-        be a square sprocket." The value returned by info() is combined with
-        other information whenever an error occurs and therefore makes more
-        sense to the user if the result is a phrase. The info() method is
-        similar in purpose and use to the **info** attribute of a validator
-        function.
-
-        Note that the result can include information specific to the particular
-        trait handler instance. If the info() method is not overridden, the
-        default method returns the value of the 'info_text' attribute.
-        """
-        return self.info_text
-
-    def get_editor(self, trait=None):
-        """ Returns a trait editor that allows the user to modify the *trait*
-        trait.
-
-        This method only needs to be specified if traits defined using this
-        trait handler require a non-default trait editor in trait user
-        interfaces. The default implementation of this method returns a trait
-        editor that allows the user to type an arbitrary string as the value.
-
-        For more information on trait user interfaces, refer to the *Traits UI
-        User Guide*.
-
-        Parameters
-        ----------
-        trait : Trait
-            The trait to be edited.
-
-        """
-        if self.editor is None:
-            self.editor = self.create_editor()
-
-        return self.editor
-
-    def create_editor(self):
-        """ Returns the default traits UI editor to use for a trait.
-        """
-        from traitsui.api import TextEditor
-
-        return TextEditor()
-
-    def inner_traits(self):
-        """ Returns a tuple containing the *inner traits* for this trait. Most
-            trait handlers do not have any inner traits, and so will return an
-            empty tuple. The exceptions are **List** and **Dict** trait types,
-            which have inner traits used to validate the values assigned to the
-            trait. For example, in *List( Int )*, the *inner traits* for
-            **List** are ( **Int**, ).
-        """
-        return ()
-
-
-# -------------------------------------------------------------------------------
-#  'TraitType' (base class for class-based trait definitions:
-# -------------------------------------------------------------------------------
-
-# Create a singleton object for use in the TraitType constructor:
-class NoDefaultSpecified(object):
-    pass
-
-
-NoDefaultSpecified = NoDefaultSpecified()
-
-
-class TraitType(BaseTraitHandler):
-    """ Base class for new trait types.
-
-        This class enables you to define new traits using a class-based
-        approach, instead of by calling the Trait() factory function with an
-        instance of a TraitHandler derived object.
-
-        When subclassing this class, you can implement one or more of the
-        method signatures below. Note that these methods are defined only as
-        comments, because the absence of method definitions in the subclass
-        definition implicitly provides information about how the trait should
-        operate.
-
-        The optional methods are as follows:
-
-        * **get ( self, object, name ):**
-
-          This is the getter method of a trait that behaves like a property.
-
-          :Parameters:
-            **object** (*object*) -- The object that the property applies to.
-
-            **name** (str) -- The name of the property on *object* property.
-
-          *Description*
-
-          If neither this method nor the set() method is defined, the value
-          of the trait is handled like a normal object attribute. If this
-          method is not defined, but the set() method is defined, the trait
-          behaves like a write-only property. This method should return the
-          value of the *name* property for the *object* object.
-
-        * **set ( self, object, name, value )**
-
-          This is the setter method of a trait that behaves like a property.
-
-          :Parameters:
-            **object** (*object*) -- The object that the property applies to.
-
-            **name** (str) -- The name of the property on *object*.
-
-            **value** -- The value being assigned as the value of the property.
-
-          *Description*
-
-          If neither this method nor the get() method is implemented, the
-          trait behaves like a normal trait attribute. If this method is not
-          defined, but the get() method is defined, the trait behaves like a
-          read-only property. This method does not need to return a value,
-          but it should raise a TraitError exception if the specified *value*
-          is not valid and cannot be coerced or adapted to a valid value.
-
-        * **validate ( self, object, name, value )**
-
-          This method validates, coerces, or adapts the specified *value* as
-          the value of the *name* trait of the *object* object. This method
-          is called when a value is assigned to an object trait that is
-          based on this subclass of *TraitType* and the class does not
-          contain a definition for either the get() or set() methods. This
-          method must return the original *value* or any suitably coerced or
-          adapted value that is a legal value for the trait. If *value* is
-          not a legal value for the trait, and cannot be coerced or adapted
-          to a legal value, the method should either raise a **TraitError** or
-          call the **error** method to raise the **TraitError** on its behalf.
-
-        * **is_valid_for ( self, value )**
-
-          As an alternative to implementing the **validate** method, you can
-          instead implement the **is_valid_for** method, which receives only
-          the *value* being assigned. It should return **True** if the value is
-          valid, and **False** otherwise.
-
-        * **value_for ( self, value )**
-
-          As another alternative to implementing the **validate** method, you
-          can instead implement the **value_for** method, which receives only
-          the *value* being assigned. It should return the validated form of
-          *value* if it is valid, or raise a **TraitError** if the value is not
-          valid.
-
-        * **post_setattr ( self, object, name, value )**
-
-          This method allows the trait to do additional processing after
-          *value* has been successfully assigned to the *name* trait of the
-          *object* object. For most traits there is no additional processing
-          that needs to be done, and this method need not be defined. It is
-          normally used for creating "shadow" (i.e., "mapped" traits), but
-          other uses may arise as well. This method does not need to return
-          a value, and should normally not raise any exceptions.
-    """
-
-    default_value = Undefined
-    metadata = {}
-
-    def __init__(self, default_value=NoDefaultSpecified, **metadata):
-        """ This constructor method is the only method normally called
-            directly by client code. It defines the trait. The
-            default implementation accepts an optional, untype-checked default
-            value, and caller-supplied trait metadata. Override this method
-            whenever a different method signature or a type-checked
-            default value is needed.
-        """
-        if default_value is not NoDefaultSpecified:
-            self.default_value = default_value
-
-        if len(metadata) > 0:
-            if len(self.metadata) > 0:
-                self._metadata = self.metadata.copy()
-                self._metadata.update(metadata)
-            else:
-                self._metadata = metadata
-            # By default, private traits are not visible.
-            if (
-                self._metadata.get("private")
-                and self._metadata.get("visible") is None
-            ):
-                self._metadata["visible"] = False
-        else:
-            self._metadata = self.metadata.copy()
-
-        self.init()
-
-    def init(self):
-        """ Allows the trait to perform any additional initialization needed.
-        """
-        pass
-
-    def get_default_value(self):
-        r"""Returns a tuple of the form: (*default_value_type*, *default_value*)
-            which describes the default value for this trait. The default
-            implementation analyzes the value of the trait's **default_value**
-            attribute and determines an appropriate *default_value_type* for
-            *default_value*. If you need to override this method to provide a
-            different result tuple, the following values are valid values for
-            *default_value_type*:
-
-                - 0, 1: The *default_value* item of the tuple is the default
-                  value.
-                - 2: The object containing the trait is the default value.
-                - 3: A new copy of the list specified by *default_value* is
-                  the default value.
-                - 4: A new copy of the dictionary specified by *default_value*
-                  is the default value.
-                - 5: A new instance of TraitListObject constructed using the
-                  *default_value* list is the default value.
-                - 6: A new instance of TraitDictObject constructed using the
-                  *default_value* dictionary is the default value.
-                - 7: *default_value* is a tuple of the form: (*callable*, *args*,
-                  *kw*), where *callable* is a callable, *args* is a tuple, and
-                  *kw* is either a dictionary or None. The default value is the
-                  result obtained by invoking callable(\*args, \*\*kw).
-                - 8: *default_value* is a callable. The default value is the
-                  result obtained by invoking *default_value*(*object*), where
-                  *object* is the object containing the trait. If the trait has
-                  a validate() method, the validate() method is also called to
-                  validate the result.
-                - 9: A new instance of TraitSetObject constructed using the
-                  *default_value* set is the default value.
-        """
-        dv = self.default_value
-        dvt = self.default_value_type
-        if dvt < 0:
-            dvt = _infer_default_value_type(dv)
-            self.default_value_type = dvt
-
-        return (dvt, dv)
-
-    def clone(self, default_value=NoDefaultSpecified, **metadata):
-        """ Clones the contents of this object into a new instance of the same
-            class, and then modifies the cloned copy using the specified
-            *default_value* and *metadata*. Returns the cloned object as the
-            result.
-
-            Note that subclasses can change the signature of this method if
-            needed, but should always call the 'super' method if possible.
-        """
-        if "parent" not in metadata:
-            metadata["parent"] = self
-
-        new = self.__class__.__new__(self.__class__)
-        new_dict = new.__dict__
-        new_dict.update(self.__dict__)
-
-        if "editor" in new_dict:
-            del new_dict["editor"]
-
-        if "_metadata" in new_dict:
-            new._metadata = new._metadata.copy()
-        else:
-            new._metadata = {}
-
-        new._metadata.update(metadata)
-
-        if default_value is not NoDefaultSpecified:
-            new.default_value = default_value
-            if self.validate is not None:
-                try:
-                    new.default_value = self.validate(
-                        None, None, default_value
-                    )
-                except:
-                    pass
-
-        return new
-
-    def get_value(self, object, name, trait=None):
-        """ Returns the current value of a property-based trait.
-        """
-        cname = TraitsCache + name
-        value = object.__dict__.get(cname, Undefined)
-        if value is Undefined:
-            if trait is None:
-                trait = object.trait(name)
-
-            object.__dict__[cname] = value = trait.default_value_for(
-                object, name
-            )
-
-        return value
-
-    def set_value(self, object, name, value):
-        """ Sets the cached value of a property-based trait and fires the
-            appropriate trait change event.
-        """
-        cname = TraitsCache + name
-        old = object.__dict__.get(cname, Undefined)
-        if value != old:
-            object.__dict__[cname] = value
-            object.trait_property_changed(name, old, value)
-
-    # -- Private Methods --------------------------------------------------------
-
-    def __call__(self, *args, **kw):
-        """ Allows a derivative trait to be defined from this one.
-        """
-        return self.clone(*args, **kw).as_ctrait()
-
-    def _is_valid_for(self, object, name, value):
-        """ Handles a simplified validator that only returns whether or not the
-            original value is valid.
-        """
-        if self.is_valid_for(value):
-            return value
-
-        self.error(object, name, value)
-
-    def _value_for(self, object, name, value):
-        """ Handles a simplified validator that only receives the value
-            argument.
-        """
-        try:
-            return self.value_for(value)
-        except TraitError:
-            self.error(object, name, value)
-
-    def as_ctrait(self):
-        """ Returns a CTrait corresponding to the trait defined by this class.
-        """
-        from .traits import CTrait
-
-        metadata = getattr(self, "_metadata", {})
-        getter = getattr(self, "get", None)
-        setter = getattr(self, "set", None)
-        if (getter is not None) or (setter is not None):
-            if getter is None:
-                getter = _write_only
-                metadata.setdefault("transient", True)
-            elif setter is None:
-                setter = _read_only
-                metadata.setdefault("transient", True)
-            trait = CTrait(TraitKind.property)
-            n = 0
-            validate = getattr(self, "validate", None)
-            if validate is not None:
-                n = _arg_count(validate)
-            trait.property(
-                getter,
-                _arg_count(getter),
-                setter,
-                _arg_count(setter),
-                validate,
-                n,
-            )
-            metadata.setdefault("type", "property")
-        else:
-            type = getattr(self, "ctrait_type", None)
-            if type is None:
-                type = trait_types.get(metadata.get("type"), 0)
-            trait = CTrait(type)
-
-            validate = getattr(self, "fast_validate", None)
-            if validate is None:
-                validate = getattr(self, "validate", None)
-                if validate is None:
-                    validate = getattr(self, "is_valid_for", None)
-                    if validate is not None:
-                        validate = self._is_valid_for
-                    else:
-                        validate = getattr(self, "value_for", None)
-                        if validate is not None:
-                            validate = self._value_for
-
-            if validate is not None:
-                trait.set_validate(validate)
-
-            post_setattr = getattr(self, "post_setattr", None)
-            if post_setattr is not None:
-                trait.post_setattr = post_setattr
-                trait.is_mapped = self.is_mapped
-
-            # Note: The use of 'rich_compare' metadata is deprecated; use
-            # 'comparison_mode' metadata instead:
-            rich_compare = metadata.get("rich_compare")
-            if rich_compare is not None:
-                trait.rich_comparison(rich_compare is True)
-
-            comparison_mode = metadata.get("comparison_mode")
-            if comparison_mode is not None:
-                trait.comparison_mode(comparison_mode)
-
-            metadata.setdefault("type", "trait")
-
-        trait.set_default_value(*self.get_default_value())
-
-        trait.handler = self
-
-        trait.__dict__ = metadata.copy()
-
-        return trait
-
-    def __getattr__(self, name):
-        if (name[:2] == "__") and (name[-2:] == "__"):
-            raise AttributeError(
-                "'%s' object has no attribute '%s'"
-                % (self.__class__.__name__, name)
-            )
-
-        return getattr(self, "_metadata", {}).get(name, None)
-
-
-# -------------------------------------------------------------------------------
-#  'TraitHandler' class (base class for all trait handlers):
-# -------------------------------------------------------------------------------
-
-
-class TraitHandler(BaseTraitHandler):
-    """ The task of this class and its subclasses is to verify the correctness
-    of values assigned to object trait attributes.
-
-    This class is an alternative to trait validator functions. A trait handler
-    has several advantages over a trait validator function, due to being an
-    object:
-
-        * Trait handlers have constructors and state. Therefore, you can use
-          them to create *parametrized types*.
-        * Trait handlers can have multiple methods, whereas validator functions
-          can have only one callable interface. This feature allows more
-          flexibility in their implementation, and allows them to handle a
-          wider range of cases, such as interactions with other components.
-
-    The only method of TraitHandler that *must* be implemented by subclasses
-    is validate().
-    """
-
-    def validate(self, object, name, value):
-        """ Verifies whether a new value assigned to a trait attribute is valid.
-
-        Parameters
-        ----------
-        object : object
-            The object whose attribute is being assigned.
-        name : str
-            The name of the attribute being assigned.
-        value :
-            The proposed new value for the attribute.
-
-        Returns
-        -------
-        If the new value is valid, this method must return either the original
-        value passed to it, or an alternate value to be assigned in place of the
-        original value. Whatever value this method returns is the actual value
-        assigned to *object.name*.
-
-        Description
-        -----------
-        This method *must* be implemented by subclasses of TraitHandler. It is
-        called whenever a new value is assigned to a trait attribute defined
-        using this trait handler.
-
-        If the value received by validate() is not valid for the trait
-        attribute, the method must called the predefined error() method to
-        raise a TraitError exception
-
-        """
-        raise TraitError(
-            "The '%s' trait of %s instance has an unknown type. "
-            "Contact the developer to correct the problem."
-            % (name, class_of(object))
-        )
 
 
 # -------------------------------------------------------------------------------
@@ -1112,27 +486,6 @@ class TraitInstance(ThisClass):
                 raise TraitError("Unable to locate class: " + args[0])
 
         return aClass(*args[1:], **kw)
-
-
-# -- Private Class --------------------------------------------------------------
-
-
-def _make_value_freed_callback(object_ref, name):
-    def _value_freed(value_ref):
-        object = object_ref()
-        if object is not None:
-            object.trait_property_changed(name, Undefined, None)
-
-    return _value_freed
-
-
-class HandleWeakRef(object):
-    def __init__(self, object, name, value):
-        object_ref = ref(object)
-        _value_freed = _make_value_freed_callback(object_ref, name)
-        self.object = object_ref
-        self.name = name
-        self.value = ref(value, _value_freed)
 
 
 # -------------------------------------------------------------------------------
@@ -1850,6 +1203,8 @@ class TraitList(TraitHandler):
 
 
 def items_event():
+    from .trait_types import Event
+
     if TraitList._items_event is None:
         TraitList._items_event = Event(
             TraitListEvent, is_base=False
@@ -1960,33 +1315,11 @@ class TraitDict(TraitHandler):
         return self.editor
 
     def items_event(self):
+        from .trait_types import Event
+
         if TraitDict._items_event is None:
             TraitDict._items_event = Event(
                 TraitDictEvent, is_base=False
             ).as_ctrait()
 
         return TraitDict._items_event
-
-
-# -------------------------------------------------------------------------------
-#  Tell the C-based traits module about the PyProtocols 'adapt' function:
-# -------------------------------------------------------------------------------
-
-from . import ctraits
-
-
-def _adapt_wrapper(*args, **kw):
-    # We need this wrapper to defer the import of 'adapt' and avoid a circular
-    # import. The ctraits 'adapt' callback needs to be set as soon as possible,
-    # but the adaptation mechanism relies on traits.
-
-    # This wrapper is called once, after which we set the ctraits callback
-    # to point directly to 'adapt'.
-
-    from traits.adaptation.api import adapt
-
-    ctraits._adapt(adapt)
-    return adapt(*args, **kw)
-
-
-ctraits._adapt(_adapt_wrapper)
