+++ traits/trait_numeric.py
@@ -27,7 +27,7 @@ import warnings
 from .constants import ComparisonMode, DefaultValue
 from .trait_base import SequenceTypes
 from .trait_errors import TraitError
-from .trait_handlers import TraitType
+from .trait_type import TraitType
 from .trait_types import Str, Any, Int as TInt, Float as TFloat
 
 # -------------------------------------------------------------------------------
